+++ clickhouse_backend/__init__.py
@@ -1,5 +1,5 @@
 from clickhouse_backend.utils.version import get_version
 
-VERSION = (1, 1, 6, "final", 0)
+VERSION = (1, 1, 7, "final", 0)
 
 __version__ = get_version(VERSION)